move rest of geo relation tests to own files
